fix: protect sync alternative routes
