+++ src/routes/sync.ts
@@ -1,22 +1,14 @@
 import { FastifyInstance } from 'fastify'
 import { syncBackfill, SyncBackfillParams } from '../lib/sync'
-import { getConfig } from '../utils/config'
-import Stripe from 'stripe'
+import { verifyApiKey } from '../utils/verifyApiKey'
 
-const config = getConfig()
+import Stripe from 'stripe'
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export default async function routes(fastify: FastifyInstance) {
   fastify.post('/sync', {
+    preHandler: [verifyApiKey],
     handler: async (request, reply) => {
-      if (!request.headers || !request.headers.authorization) {
-        return reply.code(401).send('Unauthorized')
-      }
-      const { authorization } = request.headers
-      if (authorization !== config.API_KEY) {
-        return reply.code(401).send('Unauthorized')
-      }
-
       const { created, object } =
         (request.body as {
           created?: Stripe.RangeQueryParam
