+++ src/routes/sync/weekly.ts
@@ -1,9 +1,11 @@
 import { FastifyInstance } from 'fastify'
 import { syncBackfill, SyncBackfillParams } from '../../lib/sync'
+import { verifyApiKey } from '../../utils/verifyApiKey'
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export default async function routes(fastify: FastifyInstance) {
   fastify.post('/weekly', {
+    preHandler: [verifyApiKey],
     handler: async (request, reply) => {
       const { object } = (request.body as { object?: string }) ?? {}
       const currentTimeInSeconds = Math.floor(Date.now() / 1000)
